requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-wtf==1.2.1
flask-talisman==1.1.0
bcrypt==4.1.2
pyotp==2.9.0
eventlet==0.36.1
pillow==10.2.0
wtforms==3.1.2
python-dotenv==1.0.0
gunicorn==22.0.0
